+++ src/Enums/OrderTransactionCustomFieldsEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PaynlPayment\Shopware6\Enums;
+
+class OrderTransactionCustomFieldsEnum
+{
+    const PAYNL_PAYMENTS = 'paynl_payments';
+
+    const APPROVAL_ID = 'approval_id';
+}