+++ src/Helper/ProcessingHelper.php
@@ -9,14 +9,15 @@ use PaynlPayment\Shopware6\Entity\PaynlTransactionEntity;
 use Shopware\Core\Checkout\Customer\CustomerEntity;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionDefinition;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionEntity;
+use Shopware\Core\Checkout\Order\OrderEntity;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStates;
-use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
 use Shopware\Core\Defaults;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\MultiFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\NotFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Sorting\FieldSorting;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
@@ -56,7 +57,8 @@ class ProcessingHelper
     }
 
     public function storePaynlTransactionData(
-        AsyncPaymentTransactionStruct $transaction,
+        OrderEntity $order,
+        OrderTransactionEntity $orderTransaction,
         SalesChannelContext $salesChannelContext,
         string $paynlTransactionId,
         ?Throwable $exception = null
@@ -64,17 +66,17 @@ class ProcessingHelper
         $shopwarePaymentMethodId = $salesChannelContext->getPaymentMethod()->getId();
         /** @var CustomerEntity $customer */
         $customer = $salesChannelContext->getCustomer();
-        $salesChannelId = $salesChannelContext->getSalesChannelId();
+        $salesChannelId = $salesChannelContext->getSalesChannel()->getId();
         $transactionData = [
             'paynlTransactionId' => $paynlTransactionId,
             'customerId' => $customer->getId(),
-            'orderId' => $transaction->getOrder()->getId(),
-            'orderTransactionId' => $transaction->getOrderTransaction()->getId(),
+            'orderId' => $order->getId(),
+            'orderTransactionId' => $orderTransaction->getId(),
             'paymentId' => $this->paynlApi->getPaynlPaymentMethodId($shopwarePaymentMethodId, $salesChannelId),
-            'amount' => $transaction->getOrder()->getAmountTotal(),
+            'amount' => $order->getAmountTotal(),
             'latestActionName' => StateMachineTransitionActions::ACTION_REOPEN,
             'currency' => $salesChannelContext->getCurrency()->getIsoCode(),
-            'orderStateId' => $transaction->getOrder()->getStateId(),
+            'orderStateId' => $order->getStateId(),
             // TODO: check sComment from shopware5 plugin
             'dispatch' => $salesChannelContext->getShippingMethod()->getId(),
             'exception' => (string)$exception,
@@ -103,8 +105,15 @@ class ProcessingHelper
             return "TRUE| The current transaction's commands are ignored due to a later completed transaction.";
         }
 
-        $this->updateTransactionStatus($paynlTransactionEntity, $paynlApiTransaction);
+        $paynlTransactionStatusCode = $this->getTransactionStatusFromPaynlApiTransaction($paynlApiTransaction);
+        $transitionName = $this->getOrderActionNameByPaynlTransactionStatusCode($paynlTransactionStatusCode);
+
+        $this->updateTransactionStatus($paynlTransactionEntity, $transitionName, $paynlTransactionStatusCode);
         $apiTransactionData = $paynlApiTransaction->getData();
+
+        if ($this->isUnprocessedTransactionState($paynlApiTransaction)) {
+            return sprintf('TRUE| No change made (%s)', $apiTransactionData['paymentDetails']['stateName']);
+        }
 
         return sprintf(
             'TRUE| Status updated to: %s (%s) orderNumber: %s',
@@ -168,8 +177,10 @@ class ProcessingHelper
         $salesChannelId = $paynlTransactionEntity->getOrder()->getSalesChannelId();
 
         $paynlApiTransaction = $this->getPaynlApiTransaction($paynlTransactionId, $salesChannelId);
+        $paynlTransactionStatusCode = $this->getTransactionStatusFromPaynlApiTransaction($paynlApiTransaction);
+        $transitionName = $this->getOrderActionNameByPaynlTransactionStatusCode($paynlTransactionStatusCode);
 
-        $this->updateTransactionStatus($paynlTransactionEntity, $paynlApiTransaction);
+        $this->updateTransactionStatus($paynlTransactionEntity, $transitionName, $paynlTransactionStatusCode);
     }
 
     /**
@@ -184,8 +195,27 @@ class ProcessingHelper
         $salesChannelId = $paynlTransactionEntity->getOrder()->getSalesChannelId();
 
         $paynlApiTransaction = $this->getPaynlApiTransaction($paynlTransactionId, $salesChannelId);
+        $paynlTransactionStatusCode = $this->getTransactionStatusFromPaynlApiTransaction($paynlApiTransaction);
+        $transitionName = $this->getOrderActionNameByPaynlTransactionStatusCode($paynlTransactionStatusCode);
 
-        $this->updateTransactionStatus($paynlTransactionEntity, $paynlApiTransaction);
+        $this->updateTransactionStatus($paynlTransactionEntity, $transitionName, $paynlTransactionStatusCode);
+    }
+
+    /**
+     * @param string $paynlTransactionId
+     * @param string $transitionName
+     * @param int $paynlTransactionStatusCode
+     * @return void
+     */
+    public function instorePaymentUpdateState(
+        string $paynlTransactionId,
+        string $transitionName,
+        int $paynlTransactionStatusCode
+    ): void {
+        /** @var PaynlTransactionEntity $transactionEntity */
+        $paynlTransactionEntity = $this->getPaynlTransactionEntityByPaynlTransactionId($paynlTransactionId);
+
+        $this->updateTransactionStatus($paynlTransactionEntity, $transitionName, $paynlTransactionStatusCode);
     }
 
     /**
@@ -274,28 +304,34 @@ class ProcessingHelper
 
     /**
      * @param PaynlTransactionEntity $paynlTransactionEntity
-     * @param ResultTransaction $paynlApiTransaction
+     * @param string $transitionName
+     * @param int $paynlTransactionStatusCode
+     * @return void
      */
     private function updateTransactionStatus(
         PaynlTransactionEntity $paynlTransactionEntity,
-        ResultTransaction $paynlApiTransaction
+        string $transitionName,
+        int $paynlTransactionStatusCode
     ): void {
-        $paynlTransactionStatusCode = $this->getTransactionStatusFromPaynlApiTransaction($paynlApiTransaction);
-        $orderTransactionTransitionName =
-            $this->getOrderActionNameByPaynlTransactionStatusCode($paynlTransactionStatusCode);
-
-        /** @var OrderTransactionEntity $orderTransaction */
         $orderTransaction = $paynlTransactionEntity->getOrderTransaction();
         $stateMachineStateId = $orderTransaction->getStateId();
+        $stateMachineId = $orderTransaction->getStateMachineState()->getStateMachineId();
+
+        $allowedTransitionsStatesCount = $this->getAllowedTransitionsStatesCount(
+            $transitionName,
+            $stateMachineId,
+            $stateMachineStateId
+        );
 
         if (
-            !empty($orderTransactionTransitionName)
-            && $orderTransactionTransitionName !== $paynlTransactionEntity->getLatestActionName()
+            !empty($transitionName)
+            && ($transitionName !== $paynlTransactionEntity->getLatestActionName())
+            && ($allowedTransitionsStatesCount > 0)
         ) {
             $orderTransactionId = $paynlTransactionEntity->get('orderTransactionId') ?: '';
             $stateMachine = $this->manageOrderTransactionStateTransition(
                 $orderTransactionId,
-                $orderTransactionTransitionName
+                $transitionName
             );
 
             /** @var StateMachineStateEntity $stateMachineStateEntity */
@@ -306,7 +342,7 @@ class ProcessingHelper
         $this->updatePaynlTransactionStatus(
             $paynlTransactionEntity->getId(),
             $paynlTransactionStatusCode,
-            $orderTransactionTransitionName,
+            $transitionName,
             $stateMachineStateId
         );
     }
@@ -329,6 +365,45 @@ class ProcessingHelper
         ]));
 
         return (bool)$this->paynlTransactionRepository->search($criteria, $context)->count();
+    }
+
+    private function getAllowedTransitionsStatesCount(
+        string $actionName,
+        string $stateMachineId,
+        string $stateMachineStateId
+    ): int {
+        $filter = (new Criteria())->addFilter(
+            new MultiFilter(
+                MultiFilter::CONNECTION_AND,
+                [
+                    new EqualsFilter('actionName', $actionName),
+                    new EqualsFilter('stateMachineId', $stateMachineId),
+                    new EqualsFilter('fromStateId', $stateMachineStateId),
+                    new NotFilter(NotFilter::CONNECTION_AND, [
+                        new EqualsFilter('toStateId', $stateMachineStateId),
+                    ])
+                ]
+            ));
+
+        $context = Context::createDefaultContext();
+
+        return $this->stateMachineTransitionRepository->search($filter, $context)->count();
+    }
+
+    /**
+     * @param ResultTransaction $paynlApiTransaction
+     * @return bool
+     */
+    private function isUnprocessedTransactionState(ResultTransaction $paynlApiTransaction): bool
+    {
+        $paynlTransactionStatusCode = $this->getTransactionStatusFromPaynlApiTransaction($paynlApiTransaction);
+        $transactionTransitionName = $this->getOrderActionNameByPaynlTransactionStatusCode($paynlTransactionStatusCode);
+
+        if (empty($transactionTransitionName)) {
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -358,7 +433,11 @@ class ProcessingHelper
     ): void {
         $updateData['id'] = $paynlTransactionId;
         $updateData['stateId'] = $paynlTransactionStatusCode;
-        $updateData['latestActionName'] = $orderTransactionTransitionName;
+
+        if (!empty($orderTransactionTransitionName)) {
+            $updateData['latestActionName'] = $orderTransactionTransitionName;
+        }
+
         if (!empty($stateMachineStateId)) {
             $updateData['orderStateId'] = $stateMachineStateId;
         }